Fix Coverity Scan issues CID 57634 and CID 57625

+++ src/ui/settingsdialog.cpp
@@ -263,7 +263,8 @@ void SettingsDialog::downloadCompleted()
             dir.rename(docsetDirName, tmpName);
             QtConcurrent::run([=] {
                 QDir d(dir);
-                d.cd(tmpName);
+                if (!d.cd(tmpName))
+                    return;
                 d.removeRecursively();
             });
         }
